Drop support for Go older than 1.18
We don't support Go older than 1.21 in the Go modules file, so we can
drop the special handling for older Go in the version package.
