--- version/info_go118.go
@@ -1,69 +1,0 @@
-// Copyright 2022 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build go1.18
-// +build go1.18
-
-package version
-
-import "runtime/debug"
-
-var (
-	computedRevision string
-	computedTags     string
-)
-
-func GetRevision() string {
-	if Revision != "" {
-		return Revision
-	}
-	return computedRevision
-}
-
-func GetTags() string {
-	return computedTags
-}
-
-func init() {
-	computedRevision, computedTags = computeRevision()
-}
-
-func computeRevision() (string, string) {
-	var (
-		rev      = "unknown"
-		tags     = "unknown"
-		modified bool
-	)
-
-	buildInfo, ok := debug.ReadBuildInfo()
-	if !ok {
-		return rev, tags
-	}
-	for _, v := range buildInfo.Settings {
-		if v.Key == "vcs.revision" {
-			rev = v.Value
-		}
-		if v.Key == "vcs.modified" {
-			if v.Value == "true" {
-				modified = true
-			}
-		}
-		if v.Key == "-tags" {
-			tags = v.Value
-		}
-	}
-	if modified {
-		return rev + "-modified", tags
-	}
-	return rev, tags
-}
